test: refactor infra organization
