+++ .gitignore
@@ -41,4 +41,4 @@ coverage
 /playwright-report/
 /blob-report/
 /playwright/.cache/
-/tests/.storage-state.json
+/tests/infra/.storage-state.json
