+++ tests/infra/content/image.ts
