+++ tests/infra/index.ts
@@ -1,0 +1,2 @@
+export { expect } from "@playwright/test";
+export { test } from "./test";
